ARM: dts: qcom: Separate EAS energy model for sdm636 and sdm660
In this kernel source code, we support devices using
the sdm636 and sdm660 SoCs, and as we explored, the
calculation results of the EAS energy model for the
sdm636 and sdm660 devices were very different.

The results of energy calculations for the EAS model
for sdm636 we take from the file sdm660.dtsi.

And the results of energy calculations for the EAS model
for sdm660 we take from CAF msm-4.14 (specifically for sdm660).

Signed-off-by: Fiqri Ardyansyah <fiqri15072019@gmail.com>